sw/pi: Properly use 0xFFFF0000 as a magic for PI Writes
The N64 may set the LSB of an address when performing
unaligned DMA from the PI. This leads to issues since
we used the LSB as a write command.

Instead of just checking the LSB, we can check the 16 MSB
of the value to be 0xFFFF0000.

This shouldn't be a problem since performing PI read/writes
on >= 0x80000000 isn't supported anyway.

+++ sw/picocart64_v1/utils.h
@@ -16,7 +16,7 @@ static inline uint32_t swap16(uint32_t value)
 	return (value << 16) | (value >> 16);
 }
 
-static inline uint32_t swap8(uint16_t value)
+static inline uint16_t swap8(uint16_t value)
 {
 	// 0x1122 => 0x2211
 	return (value << 8) | (value >> 8);
